feat(services): extract collection-completion bells into @mulmoclaude/collection-watchers
Fifth packages/services/* package — the collection-completion bell subsystem (fs.watch
plumbing + the convergent reconciler) shared by MulmoClaude and MulmoTerminal, the
notifier's delivery counterpart. Both hosts now run ONE watcher/reconciler over their
own collections instead of MulmoTerminal reimplementing it.

The reconcile ALGORITHM and watcher plumbing are host-agnostic; what's host-specific is
the notification TAXONOMY (which plugin namespace, priority→severity mapping) and the
in-app ROUTING (the deep-link a bell row navigates to) — the plan's "delivery sink is
MT-specific" note. Those are injected via a CollectionNotificationAdapter; the package
owns the internal slug:itemId key and round-trips it through the adapter's pluginData.

The package depends DIRECTLY on the shared singletons — @mulmoclaude/notifier (same
module instance the host configured, so publish/clear/listAll hit the same engine) and
@mulmoclaude/collection-plugin{,/server} — so only the taxonomy/routing crosses the
injection seam.

- packages/services/collection-watchers: config.ts (adapter + logger injection),
  reconciler.ts (convergent reconcile: ensure/clear/sweep, severity reconcile,
  per-key in-flight lock), watcher.ts (one fs.watch per collection + rediscovery +
  trigger-tick + single-flight). Server-only, dual ESM+CJS, peer-deps notifier +
  collection-plugin. 5 package tests drive a NON-MulmoClaude adapter through
  publish/idempotent-republish/done-clear/delete-clear to prove the seam is real.

- Host server/workspace/collections/notifications.ts reduced to a thin binding: builds
  MulmoClaude's adapter (publish under legacy "todo", legacyPriorityToSeverity,
  /collections/<slug>?selected=<itemId> via legacyActionToNavigateTarget,
  LegacyNotifierPluginData), configures the package at module load, and re-exports the
  reconciler surface. watcher.ts reduced to a re-export shim that imports notifications
  for the configure side-effect before server boot calls startCollectionWatchers.

Verification: typecheck:server/test/vue clean; the original collection suites —
notifications (40) + watcher (6) — all green against the extracted package + adapter;
package tests (5) green.
